pr: implement suggestions
