+++ library/src/test/kotlin/it/krzeminski/githubactions/yaml/StepsToYamlTest.kt
@@ -3,7 +3,6 @@ package it.krzeminski.githubactions.yaml
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import it.krzeminski.githubactions.actions.Action
-import it.krzeminski.githubactions.actions.CustomAction
 import it.krzeminski.githubactions.actions.actions.CheckoutV3
 import it.krzeminski.githubactions.actions.actions.CheckoutV3.FetchDepth
 import it.krzeminski.githubactions.actions.actions.UploadArtifactV2
@@ -35,45 +34,7 @@ class StepsToYamlTest : DescribeSpec({
                          |  run: echo 'test!'
                          |- id: someId
                          |  name: Some external action
-                         |  uses: actions/checkout@v3
-                         """.trimMargin()
-    }
-
-    describe("custom action") {
-        // given
-        val customAction = CustomAction(
-            actionOwner = "xu-cheng",
-            actionName = "latex-action",
-            actionVersion = "v2",
-            inputs = linkedMapOf(
-                "root_file" to "report.tex",
-                "compiler" to "latexmk",
-            )
-        )
-        val steps = listOf(
-            ExternalActionStep(
-                id = "latex",
-                name = "Latex",
-                action = customAction
-            )
-        )
-
-        // when
-        val yaml = steps.stepsToYaml()
-
-        // then
-        yaml shouldBe """|- id: latex
-                         |  name: Latex
-                         |  uses: xu-cheng/latex-action@v2
-                         |  with:
-                         |    root_file: report.tex
-                         |    compiler: latexmk""".trimMargin()
-
-        // given
-        val outputs = customAction.buildOutputObject("someStepId")
-
-        // when & then
-        outputs["custom-output"] shouldBe "steps.someStepId.outputs.custom-output"
+                         |  uses: actions/checkout@v3""".trimMargin()
     }
 
     describe("command step") {
